rename libE_directory.py to output_directory.py
